requirements.txt
# torch==2.8.0
# torchaudio==2.8.0
# torchmetrics==1.8.2
# torchvision==0.23.0
accelerate==0.33.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
attn_gym==0.0.4
attrs==24.2.0
bitsandbytes==0.48.1
blake3==1.0.4
boto==2.49.0
cachetools==5.5.2
cbor2==5.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorxs==1.0.1
compressed-tensors==0.11.0
contourpy==1.3.1
cryptography==43.0.3
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==3.2.0
deepspeed==0.18.0
Deprecated==1.2.18
depyf==0.19.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
easydict==1.13
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.16.1
flash_attn==2.8.3
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
genson==1.3.0
gguf==0.16.3
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.0.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.11.0
jload==0.7.1
jsave==1.6.4
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lark==1.2.2
lightning @ https://github.com/Lightning-AI/lightning/archive/refs/heads/master.zip #sha256=0bdb2b63e48ba7131f1e3a69814832721d6f56cbbb0e2beeadef62768169c770
lightning-utilities==0.15.2
listtypes==1.0.1
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.11.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.3.0
openai-harmony==0.0.4
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.7
outlines==1.2.7
outlines_core==0.2.11
pandas==1.4.3
partial-json-parser==0.2.1.1.post4
peft==0.17.1
pillow==10.4.0
ply==3.11
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==4.25.7
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.2
pydantic-extra-types==2.10.6
pydantic_core==2.41.4
pyparsing==3.2.1
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytz==2022.1
PyYAML==6.0.2
ray==2.50.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
rpds-py==0.22.3
safetensors==0.4.5
scipy==1.15.1
sentencepiece==0.2.0
sentry-sdk==2.42.1
setproctitle==1.3.7
shellingham==1.5.4
smmap==5.0.2
smt-portfolio==1.0.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.41.3
sympy==1.14.0
tabulate==0.9.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
typer==0.15.3
typing-inspection==0.4.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
vllm==0.11.0
wandb==0.22.2
watchfiles==1.0.3
websockets==14.1
wrapt==1.17.2
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
